Changed Screens + Word loading and input working
I changed the screen back to a simple rectangle screen instead of a TV frame. Client said that they wanted simple features.

The word loading system works. Right now a list of words is iterated. Each item in the list has its own word length and duration on screen calculated. This can also be changed with the loadspeed parameter.

Right now, the basic entry of words works, and the basic feedback system is being developed:
If the correct word is typed in order, then a pop-up or flash indicating "correct" should show up. And opposite for when wrong word is given.