Save&Load system +more menu's
I've got the saving and loading system working. At the moment it just saves the first three custom lists of words.
Players must click on save in the main menu for it to work.
Loading happens automatically, when the game starts up.
There's also a "clear"/"reset" button which deletes all the data from the custom lists.

I added more menus such as:
- custom list editor (which allows user to add or remove words from a custom list)
- forgot other.

I've also created a Globals.gd script which acts as the autoloader and keeps track of the current list to be used in the actual Word Testing scene