Added SFX  + working output display + change to next next round
I added sound effects to the game since everything was so silent. I also added them for buttons and selecting items, so that it confirms that buttons were pressed, instead of it just being silent.

The feedback system works mostly. When the player types a correct or incorrect answer, then either a "correct" sound will play and the correct text will go into the output box, or a "wrong" buzzer sound will play.

Once the level is finished, players have the option to either go to next level (if they have more correct answers than incorrect)
OR
option to restart the round (if have more incorrect than correct)

Lastly, I also added a "perfect round" system which displays a fireball and text when the player gets everything correct in the round. This perfect-round counter should also be in the globals script, which means the user can keep track of how many perfect rounds they have.