Countdown timer + more fixes
I added a countdown timer. It would take the sum of the durations of the text that would be on screen, as the start time. When timer reached 0, the round would be finished. This results in an automatic loss.

I also tinkered with variables like the "level increase factor" which would determine by how much the total durations would increase, and thus how "difficult" every new round will be.

---
Fixes:
- durations not being calculated every new round [ they are being calculated now]
- the condtions for winning by the checkCounter() functions. If player has more correct than incorrect WHILE timer isn't up, then that would be a win. Everything else is loss.