Changed filesave system + Activated the three predefined sets + bug fix with typuing
I changed the save file system. I just added three more variables to be added to the save file in txt format. Theya re the three predefined lists (setPhysics, setMath, SetCS). This does mean however that previous savefiles are now incompatible.

I also therefore enabled the buttons in the play menu and allowed the user to change the predefined words.

Lastly, there was bug where typing while moving the mouse would result in an error with the whole event handler. I just made it so that the get_viewport.set_inputashandled() method runs when the key is enter and when there is no mouse movement