Bug prevents + button added
I've added a bug preventer: whenever the user clicks confirm on an empty list, the game will display an OS alert message that two ore more elements are needed.
This'll prevent the game from continuing to the testing screen with only one or no items.

I also added a main menu redirector button for when the player has finished a level. This shows regardless of whether user wins or loses.