Open save folde, shuffle list, more customization
I finally managed to get Godot to open the user data directory. I added a button in the settings menu that opens the savefile folder. Now the save file can be located with one click, and then copied and backed-up (by the user of course)

I also set it up, so that the current list that's being tested gets shuffled every new level.

The client also asked for feature where the user can set a limit on how many elements should be tested.
[ E.g. There's 20 elements, but user wants to be tested on only 5 ]
Now there's a slider in the list/set editor, where you can choose how many of the words will be featured. There's a minimum of 2 elements, and the max is how long the original list is.