Started developing: StyleRank manager
I thought about adding a style rank, similar to the ones in DevilMayCry series, to this project.
It'll be one of the features to help "gamify" the memory tester experience.

It's not fully complete yet.

I'm using a Control node as the parent, with labels as children for the: style title (one letter), style rank, and animation player.
When a correct word is typed, then the rank increases. if incorrect is typed, rank decreases.

Will expand further